BACKLOG-11925: Move PickerDialog From Design System
